pyproject.toml
[project]
name = "elevenlabs-music"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.27.0",
    "fastapi>=0.127.0",
    "httpx>=0.27.0",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "openai-agents>=0.6.4",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
